Fix default discovery response endpoint logic (#382)
Conforms the RawEntityDescriptor-based logic with SAML 2.0 2.2.3:

> Complex Type IndexedEndpointType
>
> In any such sequence of like endpoints based on this type, the
> default endpoint is the first such endpoint with the isDefault
> attribute set to true. If no such endpoints exist, the default
> endpoint is the first such endpoint without the isDefault
> attribute set to false. If no such endpoints exist, the default
> endpoint is the first element in the sequence.

This was bugged because for RawEntityDescriptors, is_default
returns nil, 'true', or 'false', so 'false' endpoints are preferred
to nil endpoints.

This commit only fixes the RawEntityDescriptors case, because
that's all that we use at AAF now. Entities from FR are still
bugged because that data model doesn't distinguish between false
and nil.