fix: UpdFromFR: sync org changes for existing descriptors
Organisation for any Descriptor (Entity, IdPSSO, SPSSO, AttrAuth) was only
getting set when creating the descriptor, not when updating.

Consequently, if an EntityDescriptor was moved to a different organisation in FR,
this change wasn't getting picked in SAML Service and was not propagated into metadata.

The organisation is set from within a code block passed to create_or_update_by_fr_id,
but the block is only used in create_by_fr_id, not update_by_fr_id.

Pass the block also to update_by_fr_id and invoke it if an update
is happening (object was found).

As the code blocks is now getting invoked for both create and update,
add a guard to certain lines that only should be invoked at create time.
The guard condition relies on newly created objects not having an `id` assigned yet.

Explicitly save the object after being updated (required, otherwise changes get discarded).

Avoid invoking the save in `test` mode, as some tests set up test objects
that do not validate and cannot be saved.