fix: allow inactive AA to have no AttributeService
The constraint in the DB model that an AA must have an AttributeService
comes from the SAML 2.0 Metadata schema.

But for an inactive AA that will not get rendered into metadata, this can be relaxed.

It is necessary because of how the sync code behaves:
* An inactive role descriptor (like AA) gets synchronised
* But an inactive endpoint (like AttributeService) does not

Therefore an AA that has all AttributeServices inactive in FR (and will be inactive as such)
will have an empty set of AttributeServices in SAML Service and would fail the constraint.

The current code granting an exemption for `new?` allowed saving "just once" - still insufficient
for any changes to an inactive AA.